move webclient package up a level
